requirements.txt
Pillow~=10.0.0
PyYAML~=6.0
aiofiles~=23.1.0
aiohttp~=3.8.4